pyproject.toml
[project]
name = "courselanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "fire>=0.7.1",
    "langchain>=1.2.12",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=2.0.0",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "langsmith>=0.7.17",
]
